+++ kubernetes-client/src/main/java/io/fabric8/kubernetes/client/KubernetesClient.java
@@ -106,6 +106,44 @@ public interface KubernetesClient extends Client {
    *
    * <p>
    *   Note: your CustomResource POJO (T in this context) must implement
+   *   {@link io.fabric8.kubernetes.api.model.Namespaced} if it is a namespace-scoped resource.
+   * </p>
+   *
+   * @param resourceType Class for CustomResource
+   * @param <T> T type represents CustomResource type. If it's a namespaced resource, it must implement
+   *           {@link io.fabric8.kubernetes.api.model.Namespaced}
+   * @return returns a MixedOperation object with which you can do basic CustomResource operations
+   */
+  <T extends HasMetadata> MixedOperation<T, KubernetesResourceList<T>, Resource<T>> customResources(Class<T> resourceType);
+
+
+  /**
+   * Typed API for managing CustomResources. You would need to provide POJOs for
+   * CustomResource into this and with it you would be able to instantiate a client
+   * specific to CustomResource.
+   *
+   * <p>
+   *   Note: your CustomResource POJO (T in this context) must implement
+   *   {@link io.fabric8.kubernetes.api.model.Namespaced} if it is a namespace-scoped resource.
+   * </p>
+   *
+   * @param resourceType Class for CustomResource
+   * @param listClass Class for list object for CustomResource
+   * @param <T> T type represents CustomResource type. If it's a namespace-scoped resource, it must implement
+   *           {@link io.fabric8.kubernetes.api.model.Namespaced}
+   * @param <L> L type represents CustomResourceList type
+   * @return returns a MixedOperation object with which you can do basic CustomResource operations
+   */
+  <T extends HasMetadata, L extends KubernetesResourceList<T>> MixedOperation<T, L, Resource<T>> customResources(Class<T> resourceType, Class<L> listClass);
+
+
+  /**
+   * Typed API for managing CustomResources. You would need to provide POJOs for
+   * CustomResource into this and with it you would be able to instantiate a client
+   * specific to CustomResource.
+   *
+   * <p>
+   *   Note: your CustomResource POJO (T in this context) must implement
    *   <a href="https://github.com/fabric8io/kubernetes-client/blob/master/kubernetes-model-generator/kubernetes-model-core/src/main/java/io/fabric8/kubernetes/api/model/Namespaced.java">
    *     io.fabric8.kubernetes.api.model.Namespaced
    *   </a> if it is a Namespaced scoped resource.